backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1
pydantic==1.10.13
openai==1.3.5
python-dotenv==1.0.0
requests==2.31.0
httpx==0.25.0